# Optimize OVAL ID replacement logic