Updated OVAL ID replacement logic to replace ID only in case it is a duplicate as supposed to the current implementation that generates new IDs for each ID in each file.

**Testing:**

Verified manually through random diff checks of merged XML generated with and without the changes. The input XML files had duplicate OVAL IDs.

**Performance:**
Runtime measured using: `time oval-xml-feed-merge oci.com.ubuntu.focal.pkg.oval.xml oci.com.ubuntu.esm-apps_focal.pkg.oval.xml  --output merged.xml --verbose`
Before optimization: 
```
378.33s user 1.17s system 562% cpu 1:07.44 total
```

After optimization: 
```
153.35s user 1.29s system 505% cpu 30.586 total
```